# perf: change order of element-wise op in edge angle update calculations